setup statusCodes.js for streamlined routing errors
